+++ drivers/misc/mediatek/teei/V1.0/tz_driver/teei_keymaster.c
@@ -128,11 +128,11 @@ int __send_keymaster_command(unsigned long share_memory_size)
 	Flush_Dcache_By_Area((unsigned long)keymaster_buff_addr, (unsigned long)keymaster_buff_addr + KEYMASTER_BUFF_SIZE);
 
 	fp_call_flag = GLSCH_HIGH;
-	n_invoke_t_drv((uint64_t *)(&smc_type), 0, 0);
+	n_invoke_t_drv((uint32_t *)(&smc_type), 0, 0);
 
 	while(smc_type == 0x54) {
 		//udelay(IRQ_DELAY);
-		nt_sched_t((uint64_t *)(&smc_type));
+		nt_sched_t((uint32_t *)(&smc_type));
 	}
 
 	return 0;
@@ -171,7 +171,7 @@ int send_keymaster_command(unsigned long share_memory_size)
 		ut_pm_mutex_unlock(&pm_mutex);
 	        up(&fdrv_lock);
                 return retVal;
-        }	
+        }
 
 	down(&fdrv_sema);
 	IMSG_DEBUG("send_keymaster_command end\n");
